mmc: sdhci: add platfrom get_max_timeout hook
Currently the max_discard_to is simply got by (1 << 27) / host->timeout_clk
which is assumed to be the maximum timeout value, however, some platforms
maximum timeout counter may not be 1 << 27, e.g. i.MX uSDHC is 1 << 28.
Thus 1 << 27 may not be correct for such platforms.

It is also possible that other platforms may have different problems.
To be flexible, we add a get_max_timeout hook to get the correct
maximum timeout value for these platforms.

Signed-off-by: Dong Aisheng <b29396@xxxxxxxxxxxxx>
Reported-by: Ed Sutter <ed.sutter@xxxxxxxxxxxxxxxxxx>